Django 3.1 compatibility (#1173)
* Remove deprecated providing_args argument to Signal()

This was deprecated in Django 3.1 and will be removed in Django 4.0:

django.utils.deprecation.RemovedInDjango40Warning: The providing_args
argument is deprecated. As it is purely documentational, it has no
replacement. If you rely on this argument as documentation, you can move
the text to a code comment or docstring.

As the text explains it can be removed safely without replacement.

* Replace django.conf.urls.url() with django.urls.path()

`url()` was added in Django 2.0, deprecated in Django 3.1, and will be
removed in Django 4.0:

RemovedInDjango40Warning: django.conf.urls.url() is deprecated in favor of django.urls.re_path().

* Add Django 3.1 support in CI, setup.py, tox.ini

* Add `tox -e isort`

Allow to quickly check isort.

Also fix the isort config to use `extra_standard_library` instead of
`known_standard_library` otherwise it resets the known standard
libraries.

Also skip isort on Python 3.5 since that uses isort 4 which is
incompatible.